add memo delete button

+++ resources/views/users/show.blade.php
@@ -13,6 +13,7 @@
                 <h2> {{ $user->nomal_wants_count }} </h2>
         </div>
 
+{{-- ReallyWant一覧 --}}
  @if (count($reallyWants) > 0)    
     <div class="ichiran"><h3>Really Want一覧</h3></div> 
         <div class="mt-4 row">
@@ -29,35 +30,33 @@
                             <p class="text-center">¥{{ number_format($reallyWant->price) }}</p>
                             {{-- Want／Not  ボタン --}}  
                             <div class="text-center">@include('want.want_button',['item' => $reallyWant])</div>
-                       </div>
-                       <div class="card-footer">
+                        </div>
+                        <div class="card-footer">
                             {{-- メモの内容表示 --}}
-                            <div class="memo text-center"><p>{{ $reallyWant->pivot->memo }}</p></div>
+                            <div class="mt-2 memo text-center"><p>{{ $reallyWant->pivot->memo }}</p></div>
                         
-                        
-                         <div class="text-center">   
-                            {!! Form::open(['route' => ['wants.destroyMemo',$reallyWant->id], 'method' => 'delete']) !!}
-                            {!! Form::submit('メモ削除', ['class' => 'btn btn-warning btn-sm']) !!}
-                            {!! Form::close() !!} 
-                         </div>
-                        
-                           
-                           @if($reallyWant->pivot->memo　=== null)
-                            {{-- メモ作成フォーム--}}
-                            {!! Form::open(['route' => ['wants.storeMemo',$reallyWant->id]]) !!}
-                            <div class="mt-4 form-group">
-                             {!! Form::textarea('memo', old('memo'), ['class' => 'form-control', 'rows' => '1']) !!}
-                             {!! Form::submit('メモ入力', ['class' => 'btn btn-info btn-block']) !!}
-                              </div>
-                             {!! Form::close() !!}
+                            @if(!empty($reallyWant->pivot->memo)) 
+                                {{-- メモ削除ボタン --}}
+                                <div class="mb-4 text-center">   
+                                    {!! Form::open(['route' => ['wants.destroyMemo',$reallyWant->id], 'method' => 'delete']) !!}
+                                    {!! Form::submit('メモ削除', ['class' => 'btn btn-info btn-sm']) !!}
+                                    {!! Form::close() !!} 
+                                </div>
+                            @else 
+                                {{-- メモ作成フォーム--}}
+                                    {!! Form::open(['route' => ['wants.storeMemo',$reallyWant->id]]) !!}
+                                <div class="form-group">
+                                    {!! Form::textarea('memo', old('memo'), ['class' => 'form-control', 'rows' => '1']) !!}
+                                    {!! Form::submit('メモ入力', ['class' => 'btn btn-info btn-block']) !!}
+                                </div>
+                                    {!! Form::close() !!}
                             @endif
-                            
-                            {{--  @include('memos.form') --}}
                         </div>
                     </div>
                 </div>
             @endforeach  
         </div>  
+        
     {{-- ReallyWant一覧ページネーションのリンク --}}
     <div class="pagination justify-content-center">
         {{ $reallyWants->links() }}
@@ -65,7 +64,8 @@
 @endif   
    
    <hr />
-   
+
+{{-- Want一覧 --}}   
 @if (count($wants) > 0)  
     <div class="ichiran"><h3>Want一覧</h3></div>
          <div class="mt-4 row">
@@ -87,28 +87,28 @@
                             {{-- メモの内容表示 --}}
                             <div class="text-center memo"><p>{{ $want->pivot->memo }}</p></div>
                             
-                            <div class="text-center">   
-                            {!! Form::open(['route' => ['wants.destroyMemo',$want->id], 'method' => 'delete']) !!}
-                            {!! Form::submit('メモ削除', ['class' => 'btn btn-warning btn-sm']) !!}
-                            {!! Form::close() !!} 
-                            </div>
-                            
-                            {{-- メモ作成フォーム--}}
-                            {!! Form::open(['route' => ['wants.storeMemo',$want->id]]) !!}
-                            <div class="mt-4 form-group">
-                             {!! Form::textarea('memo', old('memo'), ['class' => 'form-control', 'rows' => '1']) !!}
-                             {!! Form::submit('メモ入力', ['class' => 'btn btn-info btn-block']) !!}
-                              </div>
-                             {!! Form::close() !!}
-                            
-                            
-                            {{-- メモ作成フォーム、削除ボタン--}}
-                           {{-- @include('memos.form')--}}
+                            @if(!empty($want->pivot->memo)) 
+                                {{-- メモ削除ボタン --}}
+                                <div class="mb-4 text-center">   
+                                    {!! Form::open(['route' => ['wants.destroyMemo',$want->id], 'method' => 'delete']) !!}
+                                    {!! Form::submit('メモ削除', ['class' => 'btn btn-info btn-sm']) !!}
+                                    {!! Form::close() !!} 
+                                </div>
+                            @else 
+                                {{-- メモ作成フォーム--}}
+                                    {!! Form::open(['route' => ['wants.storeMemo',$want->id]]) !!}
+                                <div class="form-group">
+                                    {!! Form::textarea('memo', old('memo'), ['class' => 'form-control', 'rows' => '1']) !!}
+                                    {!! Form::submit('メモ入力', ['class' => 'btn btn-info btn-block']) !!}
+                                </div>
+                                    {!! Form::close() !!}
+                            @endif
                         </div>        
                     </div>
                 </div>
             @endforeach  
          </div>  
+         
     {{-- Want一覧ページネーションのリンク --}}
     <div class="pagination justify-content-center">
         {{ $wants->links() }}
